email-priority-agent/requirements.txt
# Core Framework
langgraph>=0.0.20
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.1

# LLM and Embeddings
openai>=1.12.0
anthropic>=0.18.0
tiktoken>=0.5.2

# Vector Database
qdrant-client>=1.7.0

# Traditional Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.25
alembic>=1.13.1

# Caching
redis>=5.0.1
hiredis>=2.3.2

# Email/SMS Processing
imapclient>=2.3.1
email-validator>=2.1.0
python-magic>=0.4.27
beautifulsoup4>=4.12.3
twilio>=8.11.1

# Web Framework
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.3
pydantic-settings>=2.1.0

# Async Support
aiohttp>=3.9.1
aiofiles>=23.2.1
asyncpg>=0.29.0

# Data Processing
pandas>=2.1.4
numpy>=1.26.3
scikit-learn>=1.3.2

# NLP and Text Processing
spacy>=3.7.2
nltk>=3.8.1
transformers>=4.36.2

# Date/Time Handling
python-dateutil>=2.8.2
pytz>=2023.3

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Monitoring and Logging
structlog>=24.1.0
prometheus-client>=0.19.0
sentry-sdk>=1.39.2

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
pytest-mock>=3.12.0
httpx>=0.26.0
faker>=22.0.0

# Development Tools
black>=23.12.1
ruff>=0.1.11
mypy>=1.8.0
pre-commit>=3.6.0

# Documentation
mkdocs>=1.5.3
mkdocs-material>=9.5.3
mkdocstrings[python]>=0.24.0

# Utilities
click>=8.1.7
rich>=13.7.0
tqdm>=4.66.1
tenacity>=8.2.3